Move `blockchain.Verifier` implementation to own package
In order to avoid import cycles, move `blockchain.Verifier`
implementation into own package.
